Split progress and worker modules into focused submodules
