+++ bitloops/src/host/inference/text_generation/tests.rs
@@ -22,7 +22,11 @@ fn test_lock() -> std::sync::MutexGuard<'static, ()> {
         .expect("text-generation test lock")
 }
 
-fn write_fake_runtime_script(script_path: &Path, timeout_marker: Option<&Path>) {
+fn write_fake_runtime_script(
+    script_path: &Path,
+    timeout_marker: Option<&Path>,
+    describe_capabilities_json: &str,
+) {
     let timeout_branch = timeout_marker
         .map(|path| {
             format!(
@@ -47,7 +51,7 @@ while IFS= read -r line; do
   request_id=$(printf '%s' "$line" | sed -n 's/.*"request_id":"\([^"]*\)".*/\1/p')
   case "$line" in
     *'"type":"describe"'*)
-      printf '{{"type":"describe","request_id":"%s","protocol_version":1,"runtime_name":"bitloops-inference","runtime_version":"0.1.0","profile_name":"summary_local","provider":{{"kind":"ollama_chat","provider_name":"ollama","model_name":"ministral-3:3b","endpoint":"http://127.0.0.1:11434","capabilities":["text","json_object"]}}}}\n' "$request_id"
+      printf '{{"type":"describe","request_id":"%s","protocol_version":1,"runtime_name":"bitloops-inference","runtime_version":"0.1.0","profile_name":"summary_local","provider":{{"kind":"ollama_chat","provider_name":"ollama","model_name":"ministral-3:3b","endpoint":"http://127.0.0.1:11434","capabilities":{describe_capabilities_json}}}}}\n' "$request_id"
       ;;
     *'"type":"shutdown"'*)
       printf '{{"type":"shutdown","request_id":"%s"}}\n' "$request_id"
@@ -60,6 +64,7 @@ while IFS= read -r line; do
 done
 "#,
             timeout_branch = timeout_branch,
+            describe_capabilities_json = describe_capabilities_json,
         ),
     )
     .expect("write fake runtime script");
@@ -101,7 +106,11 @@ fn runtime_service_restarts_after_request_timeout() {
     let timeout_marker = temp.path().join("first-request-timed-out");
     let config_path = temp.path().join("bitloops.toml");
     fs::write(&config_path, "[inference]\n").expect("write fake config");
-    write_fake_runtime_script(&script_path, Some(&timeout_marker));
+    write_fake_runtime_script(
+        &script_path,
+        Some(&timeout_marker),
+        "[\"text\",\"json_object\"]",
+    );
 
     let runtime = fake_runtime_config(&script_path, &launch_log);
     let service = BitloopsInferenceTextGenerationService::new(
@@ -137,7 +146,7 @@ fn runtime_service_reuses_hot_runtime_across_service_instances() {
     let launch_log = temp.path().join("launches.log");
     let config_path = temp.path().join("bitloops.toml");
     fs::write(&config_path, "[inference]\n").expect("write fake config");
-    write_fake_runtime_script(&script_path, None);
+    write_fake_runtime_script(&script_path, None, "[\"text\",\"json_object\"]");
 
     let runtime = fake_runtime_config(&script_path, &launch_log);
     let first = BitloopsInferenceTextGenerationService::new(
@@ -193,7 +202,68 @@ fn runtime_service_shuts_down_after_idle_eviction() {
     let launch_log = temp.path().join("launches.log");
     let config_path = temp.path().join("bitloops.toml");
     fs::write(&config_path, "[inference]\n").expect("write fake config");
-    write_fake_runtime_script(&script_path, None);
+    write_fake_runtime_script(&script_path, None, "[\"text\",\"json_object\"]");
+
+    let runtime = fake_runtime_config(&script_path, &launch_log);
+    let first = BitloopsInferenceTextGenerationService::new(
+        "summary_local",
+        "ollama_chat",
+        &runtime,
+        &config_path,
+    )
+    .expect("build first runtime service");
+    assert_eq!(
+        serde_json::from_str::<Value>(&first.complete("system", "user").expect("first completion"))
+            .expect("parse first completion"),
+        serde_json::json!({
+            "summary": "Summarises the symbol.",
+            "confidence": 0.91
+        })
+    );
+
+    evict_idle_text_generation_sessions_for_tests(Duration::ZERO);
+
+    let second = BitloopsInferenceTextGenerationService::new(
+        "summary_local",
+        "ollama_chat",
+        &runtime,
+        &config_path,
+    )
+    .expect("build second runtime service");
+    assert_eq!(
+        serde_json::from_str::<Value>(
+            &second
+                .complete("system", "user")
+                .expect("second completion")
+        )
+        .expect("parse second completion"),
+        serde_json::json!({
+            "summary": "Summarises the symbol.",
+            "confidence": 0.91
+        })
+    );
+
+    let launches = fs::read_to_string(&launch_log).expect("read launch log");
+    assert_eq!(
+        launches.lines().count(),
+        2,
+        "expected a new runtime launch after idle eviction, got: {launches}"
+    );
+}
+
+#[test]
+fn runtime_service_accepts_structured_provider_capabilities_from_describe() {
+    let _guard = test_lock();
+    let temp = TempDir::new().expect("temp dir");
+    let script_path = temp.path().join("fake_inference_runtime.sh");
+    let launch_log = temp.path().join("launches.log");
+    let config_path = temp.path().join("bitloops.toml");
+    fs::write(&config_path, "[inference]\n").expect("write fake config");
+    write_fake_runtime_script(
+        &script_path,
+        None,
+        r#"{"response_modes":["text","json_object"],"usage_reporting":true}"#,
+    );
 
     let runtime = fake_runtime_config(&script_path, &launch_log);
     let first = BitloopsInferenceTextGenerationService::new(
